==> notebook/1.jsonToDataFrame.ipynb <==
import sys
sys.path.append("../src")

import pandas as pd
from dataLoad import load_json

json_item_names_table = pd.read_json("../dataset/item_names_table.json", orient="index")
item_names_table_df = json_item_names_table.reset_index().rename(columns={"index": "item_id", 0: "item_name"})
item_names_table_df["item_id"] = item_names_table_df["item_id"].str.replace("item_", "").astype(int)

display(item_names_table_df)
item_names_table_df.to_parquet("../dataset/item_names_table.parquet")

json_test_user_label = pd.read_json("../dataset/test_user_label.json", orient="index")
test_user_label_df = json_test_user_label.reset_index().rename(columns={"index": "user", 0: "item_id", 1: "timestamp"})
test_user_label_df["timestamp"] = pd.to_datetime(test_user_label_df["timestamp"])
test_user_label_df["user"] = test_user_label_df["user"].str.replace("user_", "").astype(int)
test_user_label_df["item_id"] = test_user_label_df["item_id"].str.replace("item_", "").astype(int)

display(test_user_label_df)
test_user_label_df.to_parquet("../dataset/test_user_label.parquet")

json_test_user_seq_log = load_json("../dataset/test_user_seq_log.json")

df_list = []
for user, items in json_test_user_seq_log.items():
    for item in items:
        if item:
            df_list.append([user, item[0], item[1]])

test_user_seq_log_df = pd.DataFrame(df_list, columns=["user", "item_id", "timestamp"])
test_user_seq_log_df["timestamp"] = pd.to_datetime(test_user_seq_log_df["timestamp"])
test_user_seq_log_df["user"] = test_user_seq_log_df["user"].str.replace("user_", "").astype(int)
test_user_seq_log_df["item_id"] = test_user_seq_log_df["item_id"].str.replace("item_", "").astype(int)

display(test_user_seq_log_df)
test_user_seq_log_df.to_parquet("../dataset/test_user_seq_log.parquet")

json_train_user_seq_log = load_json("../dataset/train_user_seq_log.json")
df_list = []
for user, items in json_train_user_seq_log.items():
    for item in items:
        if item:
            df_list.append([user, item[0], item[1]])

train_user_seq_log_df = pd.DataFrame(df_list, columns=["user", "item_id", "timestamp"])
train_user_seq_log_df["timestamp"] = pd.to_datetime(train_user_seq_log_df["timestamp"])
train_user_seq_log_df["user"] = train_user_seq_log_df["user"].str.replace("user_", "").astype(int)
train_user_seq_log_df["item_id"] = train_user_seq_log_df["item_id"].str.replace("item_", "").astype(int)

display(train_user_seq_log_df)
train_user_seq_log_df.to_parquet("../dataset/train_user_seq_log.parquet")
