==> notebook/3.featureExtract.ipynb <==
import sys
sys.path.append("../src")
import pandas as pd
from dataLoad import dataLoad
from datetime import datetime

path= "../dataset/"
item_names_table, train_user_seq_log, test_user_label, test_user_seq_log = dataLoad(path)

1. 요일(0~6)
2. 정수 매핑 일자 (0~90)
3. 시간 (0~23)
4. 정수 매핑 주차 (1~13)
5. mp : 아이템 인기도(누적 클릭 횟수)
- 고민해볼 문제 : 클릭이 한번도 안 된 아이템은 학습에서 제거할 것인가? : O    
6. 유저의 아이템 클릭 횟수 (얼마나 많은 로그를 남기는 충성고객인지)
7. 아이템 카테고리 (자연어 처리 필요)
8. 아이템 브랜드 (자연어 처리 필요)

train_timestamp = train_user_seq_log["timestamp"]
train_timestamp

# 1. 요일(1~7)

train_timestamp_df = train_timestamp.to_frame(name="timestamp")
train_user_seq_log["day_of_week"] = train_timestamp_df["timestamp"].dt.dayofweek

# 2. 정수 매핑 일자 (1~31)

reference_date = datetime(2021, 2, 11)
days = (train_timestamp - reference_date).dt.total_seconds() / (24 * 3600)
days_df = days.astype(int).to_frame(name="days")

train_user_seq_log = pd.concat([train_user_seq_log, days_df], axis=1)

# 3. 시간 (1~24)
train_user_seq_log["hour"] = train_timestamp_df["timestamp"].dt.hour

# 4. 정수 매핑 주차 (1~13)

weeks = (days / 7).astype(int) + 1
weeks_df = weeks.to_frame(name="weeks")
train_user_seq_log = pd.concat([train_user_seq_log, weeks_df], axis=1)

# 5. mp : 아이템 인기도(누적 클릭 횟수)
train_user_seq_log = train_user_seq_log.drop_duplicates() 
# 5-1 . 누적집계
# 시퀀셜 데이터이므로 시간에 따른 누적량 증가 확인 가능.
cumcount = train_user_seq_log.groupby("item_id").cumcount().to_frame(name="cumcount")
cumcount["cumcount"] = cumcount["cumcount"] + 1
train_user_seq_log = pd.concat([train_user_seq_log, cumcount], axis=1)

# 5-2 . 90일간의 총 누적집계 (mp ranking)
train_user_seq_log_df = train_user_seq_log.copy()
mp_item = train_user_seq_log_df.item_id.value_counts().reset_index().rename(columns={"index":"item_id","item_id":"click_count"})
### normalize column
min_val = mp_item["click_count"].min()
range_val = mp_item["click_count"].max() - mp_item["click_count"].min()
mp_item["click_count_normalized"] = (mp_item["click_count"] - min_val) / range_val
train_user_seq_log = pd.merge(train_user_seq_log, mp_item, on="item_id", how="inner")

# 6. 유저 충성도
train_user_seq_log_df = train_user_seq_log.copy()
user_click = train_user_seq_log_df.user.value_counts().reset_index().rename(columns={"index":"user","user":"user_click_count"})
### normalize column
min_val = user_click["user_click_count"].min()
range_val = user_click["user_click_count"].max() - user_click["user_click_count"].min()
user_click["user_click_count_normalized"] = (user_click["user_click_count"] - min_val) / range_val
train_user_seq_log = pd.merge(train_user_seq_log, user_click, on="user", how="inner")

# 7. 아이템 카테고리 (자연어 처리 필요)

# 일단은 여기까지 하고, 만든 임베딩을 넣어 모델 돌리기 시도

# 8. 아이템 브랜드 (자연어 처리 필요)

train_user_seq_log

train_user_seq_log.to_parquet("../dataset/fe_train_user_seq_log.parquet")

# featureEngineering.py에 위 코드 정리 완료. 실행 시 fe_ 버전으로 train 데이터셋 반환.
# 다만 notebook들에서는 parquet datatype로 미리 저장해둔 file을 dataLoad.py에서 바로 불러올 수 있도록 함
# 촉박한 시간 관계상 shortcut 활용
