==> notebook/8.LgbmRanker.ipynb <==
import sys
sys.path.append("../src")
import pandas as pd
import numpy as np
from tqdm import tqdm
import pickle
import matplotlib.pyplot as plt
import lightgbm
from dataLoad import trainValidLoad
from metric import ndcg_calculator, hit_at_k

path= "../dataset/"
train, train_valid, sample_sumbission = trainValidLoad(path)

train = train.sort_values(by ="timestamp", ascending = True)\
        .reset_index(drop=True).drop(columns=["timestamp","click_count_normalized","user_click_count_normalized"])

# 현 데이터셋에서는 train에서 또다시 valid를 나눌 수가 없다.
# 이미 train에 단 1건의 데이터만 가진 유저가 존재하기 때문.
# 그러니 우리는 train_valid의 기록을 가지고 label_df를 만들어야 한다. 

train_user_click_count =train.user.value_counts()
print(f"유저 별 아이템 클릭수 최소: {train_user_click_count.min()}")

def make_label(train_valid):
    label_df = train_valid[["user","item_id"]]
    label_df.drop_duplicates(subset=["user","item_id"],inplace=True)
    label_df["click_num"] = 1
    return label_df

def label_create_df(train, label_df):
    train = pd.merge(train, label_df, how="left", on=["user","item_id"])
    train["click_num"] = train["click_num"].fillna(0)
    return train

label_df = make_label(train_valid)
train = label_create_df(train, label_df)

# print(train[train["click_num"]==1].count()) #91997
# print(train[train["click_num"]==0].count()) #16298892

# cadidate 생성 필요
# train with basic featuresExtract

def _ndcg_calculator(gt, rec, idcg):
    if not isinstance(rec, list):
        rec = [rec]
    dcg = 0.0
    for i, r in enumerate(rec):
        if r in set(gt):
            dcg += 1.0 / np.log(i + 2)
    return dcg / idcg

def ndcg_calculator(answer, submission, n=10):
    idcg = sum((1.0 / np.log(i + 1) for i in range(1, n + 1)))
    assert (answer.user != submission.user).sum() == 0
    ndcg_list = []
    for (_, row_answer), (_, row_submit) in zip(answer.iterrows(), submission.iterrows()):
        ndcg_list.append(_ndcg_calculator(row_answer.item_id, row_submit.predicted_list, idcg))
    ndcg_score = sum(ndcg_list) / len(answer)
    return ndcg_score


def hit_at_k(answer, submission, n=10):
    assert (answer.user != submission.user).sum() == 0
    if n > len(answer):
        n = len(answer)
    hit = 0
    for i in range(len(answer)):
        answer_ids = answer["item_id"].loc[i]
        submission_ids = submission["predicted_list"].loc[i]
        hits = np.isin(submission_ids, answer_ids)
        hit += np.sum(hits) / len(hits)
    hit /= len(answer)
    return hit


# 시간이 된다면 class 리팩토링
def lgbm_preprocess(train):
    X_train = train.drop(columns=["click_num"])
    y_train = train["click_num"]
#     group_dict = train.groupby("user")["user"].count().to_dict()
    item_idx = X_train["item_id"].copy()
    user_idx = X_train["user"].copy()
    del X_train["item_id"], X_train["user"]
    return X_train, y_train, item_idx, user_idx

def make_groups(X_train):
    group_dict = []
    batchsize = len(X_train)
    while True:
        if batchsize >= 2000:
            group_dict.append(2000)
            batchsize = batchsize - 2000
        else:
            group_dict.append(batchsize)
            break
            
def train_LgbmRanker_batch(X_train, y_train, model_params, batch_size):
    model = lightgbm.LGBMRanker(
        objective="lambdarank",
        metric="ndcg",
        boosting_type="dart",
        num_leaves= model_params["num_leaves"],
        learning_rate=model_params["learning_rate"],
        n_estimators= model_params["n_estimators"],
        importance_type="gain",
        verbose= -1,
        random_state= 42
        )
    num_batches = X_train.shape[0] // batch_size
    for i in tqdm(range(0, X_train.shape[0], batch_size), total=num_batches):
        X_batch = X_train[i:i + batch_size]
        y_batch = y_train[i:i + batch_size]
#         user_array = X_batch.user.to_numpy()
#         group_vectorized = np.vectorize(lambda x: group_dict.get(x, 0))
#         train_group_batch = group_vectorized(user_array) 
        group_dict = []
        batchsize = len(X_batch)
        while batchsize !=0 :
            if batchsize >= 1000:
                group_dict.append(1000)
                batchsize = batchsize - 1000
            else:
                group_dict.append(batchsize)
                batchsize -= batchsize 
        model.fit(
        X=X_batch,
        y=y_batch,
        group=group_dict,
        ) 
    feature_importances_df = pd.DataFrame(dict(zip(X_train.columns, model.feature_importances_)), index=["feature_importances"]).T
    return model, feature_importances_df

    
def valid_evaluation(X_train, train_valid, sample_sumbission, model, feature_importances_df, item_idx, user_idx): 
    print(feature_importances_df)
    pred = model.predict(X_train)
    X_train["pred"] = pred
    X_train["item_id"] = item_idx
    X_train["user"] = user_idx
    
    print("performance")
    # each user pred 25 items
    lgbm_sub_df = X_train.sort_values(by="pred", ascending=False).groupby("user").head(25)
    lgbm_user_items_dict = lgbm_sub_df.groupby("user")["item_id"].unique().to_dict()
    sample_sumbission["predicted_list"] = sample_sumbission["user"].apply(lambda x: lgbm_user_items_dict.get(x, []))

    print("lgbm ndcg:", ndcg_calculator(sample_sumbission, train_valid, 10))
        
    return X_train, sample_sumbission

import gc
gc.collect()

model_params={"num_leaves":150 ,"learning_rate":0.005,"n_estimators":35}

X_train, y_train, item_idx, user_idx = lgbm_preprocess(train)

model, feature_importances_df = train_LgbmRanker_batch(X_train, y_train, model_params, 10000)

sample_sumbission = sample_sumbission.drop(columns=["predicted_list"])
sample_sumbission= sample_sumbission.rename(columns={"item_id":"predicted_list"})

sample_sumbission
# 부족한 row는 앙상블해서 채우는 작업 필요 

X_train, sample_sumbission = \
valid_evaluation(X_train, train_valid, sample_sumbission, model, feature_importances_df, item_idx, user_idx, 10)
#  nDCG@10

%%time
mf_hit = hit_at_k(train_valid, sample_sumbission, 10)

print("performance")
print(f"hit@10(mf_ALS): {mf_hit:.4f}")

%%time
mf_hit = hit_at_k(train_valid, sample_sumbission, 1)

print("performance")
print(f"hit@1(mf_ALS): {mf_hit:.4f}")

X_train.to_parquet("../output/batch1000_lgbm_X_train.parquet", index=False)
pickle.dump(model, open("../output/batch1000_lgbm_model.pkl", "wb"))
