==> notebook/11.BidirectionalLSTM.ipynb <==
import sys
sys.path.append("../src")
from dataLoad import dataLoad, fe_trainLoad

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from tqdm import tqdm
import gc
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras import layers
from sklearn.model_selection import train_test_split


path= "../dataset/"
item_names_table, train_user_seq_log, test_user_label, test_user_seq_log = dataLoad(path)
fe_train_user_seq_log = fe_trainLoad(path)

COLUMNS = [
    "user",
    "item_id",
    "day_of_week",
    "days",
    "hour",
    "weeks",
    # "cumcount",
    # "click_count_normalized",
    # "user_click_count_normalized"
]
N_SAMPLES = 500000

transactions = fe_train_user_seq_log.iloc[:N_SAMPLES]#.merge(item_names_table, on = ("item_id"))[COLUMNS]

transactions.item_id.nunique()

ids = np.arange(transactions.item_id.nunique())
# shuffling to prevent any correlation between new labels and outcome
np.random.shuffle(ids)

# mapping article id to integer encoding 
map_item_id = {i:j for i, j in zip(transactions.item_id.unique(), ids)}

transactions["item_map"] = transactions["item_id"].map(map_item_id)

transactions.head()

transactions_train = transactions.groupby("user").agg({"item_map":lambda x: list(x)})

transactions_train.head()

# Padding to make all lists of same size 
length = max(map(len, transactions_train.item_map))
X = np.asarray([[0]*(length-len(xi)) + xi for xi in transactions_train.item_map]).astype("int32")
X

X_train, X_val, y_label_train, y_val_train = train_test_split(X[:,:-1], X[:, -1], test_size=0.2, shuffle=True, random_state=42)
y_train = tf.keras.utils.to_categorical(y_label_train, num_classes=transactions.item_id.nunique())
# onehot encoding on labels corresponding to num_classes
y_val = tf.keras.utils.to_categorical(y_val_train, num_classes=transactions.item_id.nunique())
X_train.shape

gc.collect()

# BASIC Bidirectional-LSTM Model Architecture

model = tf.keras.Sequential()
# Adding an Embbeding Layer
model.add(layers.Embedding(transactions.item_id.nunique(), 20, input_length=X_train.shape[1]))
model.add(layers.Bidirectional(layers.LSTM(64, dropout=0.2)))
model.add(layers.Dense(transactions.item_id.nunique(), activation="softmax"))

model.summary()

from tensorflow.keras.optimizers import RMSprop, Adam

optimizer = Adam(learning_rate=0.01)
model.compile(loss="categorical_crossentropy", optimizer=optimizer, metrics=["accuracy"])

history = model.fit(X_train, y_train, epochs=3, validation_data=(X_val, y_val), verbose=2, batch_size=64)
model.save("./model1.h5")

fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(15, 7))
ax1.plot(history.history["accuracy"])
ax2.plot(history.history["loss"], color="Orange")
# plt.plot(history.history["val_acc"])
# plt.title("model accuracy")
# plt.ylabel("accuracy")
ax2.set_xlabel("epoch")
ax1.set_ylabel("accuracy")
ax2.set_ylabel("loss")
ax1.set_title("Train Accuracy/Epoch")
ax2.set_title("Train Loss/Epoch")


plt.show()
